+++ examples/timer.rb
@@ -19,12 +19,8 @@ class Timer
 
   def stop_alarm
     if @pid
-      if @th.alive?
-        Process.kill(:SIGKILL, @pid)
-        @pid = nil
-      else
-        @pid = nil
-      end
+      Process.kill(:SIGKILL, @pid) if @th.alive?
+      @pid = nil
     end
   end
